+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed the broken footnote in-place preview and editing support when using named references.
 - Improved the design of wrongly spelled words -- now the dotted line is closer to the actual words.
 - Fixed `Alt`-clicking files in the combined sidebar mode. Now this will also open QuickLooks.
+- Added the shortcuts `Cmd/Ctrl+Shift+E` to focus the editor and `Cmd/Ctrl+Shift+T` to focus the file list.
 
 ## Under the Hood